Implemented the rest of Reed-Solomon error correction
This is now passing both the test_correction and test_detection tests,
which should be providing decent test coverage now.

There's likely a lot of room left for improvement. Do we need all these
polynomial buffers? Reuse a bit better? Do we really need the constant
term in our error-locators?

But this is at least a pretty good proof-of-concept.