Split the static math_buffer back into separate polynomial buffers
Yeah, I realize this is a big pain to configure in malloc-less
environments, but it does give the linker a better chance to place
things optimally.

The previous math_buffer being 1*code_size + 4*ecc_size would be rather
awkward for a linker. Both code_size and ecc_size would often be a
power-of-two, but the math_buffer probably not...