Added optional ROM-backed p config option
I thought this would be more effort than it was worth. But after putting
together rs-poly.py realized the only remaining bit of work was a config
option and a bit of documentation.

Added test_rom to test that this works as intended.

---

One downside is we still pay the code cost for generating the generator
polynomial, even if provided statically, but this is just an example.

We're limited a bit by sticking to littlefs's config patterns, which
need some significant improvements...